Issue #351: [cocoa] Refactoring: extract keyboard related functions
This commit seeks to extract some functions without changing the code.

The only change is that `Display.post()` now correctly releases the
result of `TISCopyCurrentKeyboardInputSource()`.

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/OS.java
@@ -300,8 +300,8 @@ public static final native void GetThemeMetric(int themeConstant, int[] metric);
 */
 public static final native int HIThemeDrawFocusRect(CGRect inRect, boolean inHasFocus, long inContext, int inOrientation);
 
-public static final int kUCKeyActionDown = 0;
-public static final int kUCKeyActionUp = 1;
+public static final short kUCKeyActionDown = 0;
+public static final short kUCKeyActionUp = 1;
 /** @method flags=dynamic */
 public static final native byte LMGetKbdType();
